API_CRAWLER/requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
astroid==2.2.5
Automat==0.7.0
beautifulsoup4==4.7.1
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.6.1
cssselect==1.0.3
CurrencyConverter==0.13.9
fake-useragent==0.1.11
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Login==0.4.1
Flask-Redis==0.3.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.19
itsdangerous==1.1.0
jmespath==0.9.4
jsmin==2.2.2
kombu==4.5.0
lazy-object-proxy==1.4.1
lxml==4.3.3
mccabe==0.6.1
nbconvert==5.5.0
nbformat==4.4.0
notebook==5.7.8
parsel==1.5.1
pep8==1.7.1
pkg-resources==0.0.0
psycopg2==2.8.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycparser==2.19
pycurl==7.43.0.2
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pylint==2.3.1
pyOpenSSL==19.0.0
pyquery==1.4.0
pyspider==0.3.10
python-dotenv==0.10.2
pytz==2019.1
PyYAML==5.1
qtconsole==4.4.4
queuelib==1.5.0
redis==3.2.1
requests==2.21.0
Scrapy==1.6.0
scrapy-splash==0.7.2
selenium==3.141.0
service-identity==18.1.0
soupsieve==1.9.1
tblib==1.3.2
Twisted==19.2.1
typed-ast==1.3.5
u-msgpack-python==2.5.1
urllib3==1.24.2
vine==1.3.0
w3lib==1.20.0
Werkzeug==0.15.2
wrapt==1.11.1
WsgiDAV==3.0.0
zope.interface==4.6.0
